feat: allow eval context to be extended with tm_vendor function (#702)
