+++ event/event.go
@@ -12,8 +12,30 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package event implements a simple event stream.
+// Package event implements a simple event stream and defines all
+// events generated by different parts of Terramate.
 package event
+
+import (
+	"github.com/mineiros-io/terramate/project"
+	"github.com/mineiros-io/terramate/tf"
+)
+
+type (
+	// TmVendorCall represents the event that is dispatched when a tm_vendor
+	// function is called.
+	TmVendorCall struct {
+		// Source it the tm_vendor source argument
+		Source string
+	}
+
+	// VendorProgress represents a vendor progress event.
+	VendorProgress struct {
+		Message   string
+		TargetDir project.Path
+		Module    tf.Source
+	}
+)
 
 // Stream is a stream of events.
 type Stream[T any] chan T
